refactor(database): extract stmt cache (stmt_cache.go) and vector utilities (vectors.go); stub originals (no behavior change)
